Build: switch plugin for generating all-in-one jar
Motivation:

The existing pom configuration uses the onejar-maven-plugin from
org.dstovall, but this appears to have be removed.

Modification:

Switch to using maven-assembly-plugin, with the jar-woth-dependencies
descriptor.  This results in maven building both a normal and a
with-dependencies jar file.

Some additional manifest items were added.

Result:

Building works as expected.

Signed-off-by: Paul Millar <paul.millar@desy.de>